# miniminishop-api-doc/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.0.1",
    "title": "Miniminishop",
    "description": "Miniminishop"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "local - server"
    },
    {
      "url": "http://localhost:8080",
      "description": "test - server"
    },
    {
      "url": "https://localhost:8080",
      "description": "test - server"
    }
  ],
  "tags": [
    {
      "name": "ApiDelivery",
      "description": "ApiDelivery Controller"
    },
    {
      "name": "ApiItem",
      "description": "ApiItem Controller"
    },
    {
      "name": "ApiMember",
      "description": "ApiMember Controller"
    },
    {
      "name": "ApiOrder",
      "description": "ApiOrder Controller"
    },
    {
      "name": "ApiOrderiem",
      "description": "ApiOrderiem Controller"
    }
  ],
  "paths": {
    "/api/v1//delivery/add": {
      "post": {
        "summary": "배달 등록",
        "operationId": "insertDelivery",
        "tags": [
          "ApiDelivery"
        ],
        "requestBody": {
          "description": "배달 등록",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertDeliveryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsertDeliveryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1//delivery/update": {
      "post": {
        "summary": "배달 상태 수정",
        "operationId": "updateDelivery",
        "tags": [
          "ApiDelivery"
        ],
        "requestBody": {
          "description": "배달 상태 수정",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDeliveryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateDeliveryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1//item/add": {
      "post": {
        "summary": "상품 추가",
        "operationId": "insertItem",
        "tags": [
          "ApiItem"
        ],
        "requestBody": {
          "description": "상품 추가",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsertItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1//item/delete": {
      "post": {
        "summary": "상품 삭제",
        "operationId": "deleteItem",
        "tags": [
          "ApiItem"
        ],
        "requestBody": {
          "description": "상품 삭제",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1//item/update": {
      "post": {
        "summary": "상품 수정",
        "operationId": "updateItem",
        "tags": [
          "ApiItem"
        ],
        "requestBody": {
          "description": "상품 수정",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1//item/list": {
      "get": {
        "summary": "상품 목록 조회",
        "operationId": "selectItemList",
        "tags": [
          "ApiItem"
        ],
        "description": "상품 목록 조회",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectItemListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1//item/getByName": {
      "get": {
        "summary": "상품 이름으로 조회",
        "operationId": "selectItemByName",
        "tags": [
          "ApiItem"
        ],
        "description": "상품 이름으로 조회",
        "parameters": [
          {
            "in": "query",
            "name": "itemName",
            "required": true,
            "description": null,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectItemByNameResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1//item/getById": {
      "get": {
        "summary": "상품 아이디로 조회",
        "operationId": "selectItemById",
        "tags": [
          "ApiItem"
        ],
        "description": "상품 아이디로 조회",
        "parameters": [
          {
            "in": "query",
            "name": "itemId",
            "required": true,
            "description": null,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectItemByIdResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1//member/add": {
      "post": {
        "summary": "회원 가입",
        "operationId": "insertMember",
        "tags": [
          "ApiMember"
        ],
        "requestBody": {
          "description": "회원 가입",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertMemberRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsertMemberResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1//member/getMemberList": {
      "get": {
        "summary": "회원 리스트 가져오기",
        "operationId": "selectMemberList",
        "tags": [
          "ApiMember"
        ],
        "description": "회원 리스트 가져오기",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectMemberListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1//member/getMemberByid": {
      "get": {
        "summary": "회원 번호로 회원 찾아오기",
        "operationId": "selectMemberById",
        "tags": [
          "ApiMember"
        ],
        "description": "회원 번호로 회원 찾아오기",
        "parameters": [
          {
            "in": "query",
            "name": "memberId",
            "required": true,
            "description": null,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectMemberByIdResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1//order/add": {
      "post": {
        "summary": "주문",
        "operationId": "insertOrder",
        "tags": [
          "ApiOrder"
        ],
        "requestBody": {
          "description": "주문",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertOrderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsertOrderResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1//order/status": {
      "post": {
        "summary": "주문 상태 변경",
        "operationId": "updateOrderStatus",
        "tags": [
          "ApiOrder"
        ],
        "requestBody": {
          "description": "주문 상태 변경",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateOrderStatusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1//orderitem/add": {
      "post": {
        "summary": "주문상품 담기",
        "operationId": "insertOrderItem",
        "tags": [
          "ApiOrderiem"
        ],
        "requestBody": {
          "description": "주문상품 담기",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertOrderItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsertOrderItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1//orderitem/remove": {
      "post": {
        "summary": "주문상품 덜기",
        "operationId": "deleteOneOrderItem",
        "tags": [
          "ApiOrderiem"
        ],
        "requestBody": {
          "description": "주문상품 덜기",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteOneOrderItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteOneOrderItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1//ordertiem/getByOrderId": {
      "get": {
        "summary": "주문상품 주문 아이디로 찾기",
        "operationId": "getOrderitemByMemberId",
        "tags": [
          "ApiOrderiem"
        ],
        "description": "주문상품 주문 아이디로 찾기",
        "parameters": [
          {
            "in": "query",
            "name": "member_id",
            "required": true,
            "description": null,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetOrderitemByMemberIdResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    },
    "schemas": {
      "Member": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "number"
          },
          "memberType": {
            "type": "integer"
          },
          "memberName": {
            "type": "string"
          },
          "memberPhone": {
            "type": "string"
          },
          "memberStartDate": {
            "type": ""
          }
        }
      },
      "Orders": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "number"
          },
          "orderStatus": {
            "type": "string"
          },
          "memberId": {
            "type": "number"
          },
          "deliveryId": {
            "type": "number"
          }
        }
      },
      "Delivery": {
        "type": "object",
        "properties": {
          "deliveryId": {
            "type": "number"
          },
          "deliveryStatus": {
            "type": "string"
          },
          "member_id": {
            "type": "number"
          },
          "memberId": {
            "type": "number"
          }
        }
      },
      "Item": {
        "type": "object",
        "properties": {
          "itemId": {
            "type": "number"
          },
          "itemName": {
            "type": "string"
          },
          "itemAmount": {
            "type": "integer"
          },
          "itemPrice": {
            "type": "number"
          }
        }
      },
      "Orderitem": {
        "type": "object",
        "properties": {
          "orderitem_id": {
            "type": "number"
          },
          "memberId": {
            "type": "number"
          },
          "itemId": {
            "type": "number"
          },
          "itemPrice": {
            "type": "number"
          },
          "itemAmount": {
            "type": "integer"
          },
          "orderId": {
            "type": "number"
          }
        }
      },
      "InsertDeliveryRequest": {
        "type": "object",
        "properties": {
          "deliveryStatus": {
            "type": "string"
          },
          "memberId": {
            "type": "number"
          }
        }
      },
      "UpdateDeliveryRequest": {
        "type": "object",
        "properties": {
          "deliveryId": {
            "type": "number"
          },
          "deliveryStatus": {
            "type": "string"
          }
        }
      },
      "InsertItemRequest": {
        "type": "object",
        "properties": {
          "itemName": {
            "type": "string"
          },
          "itemAmount": {
            "type": "integer"
          },
          "itemPrice": {
            "type": "number"
          }
        }
      },
      "DeleteItemRequest": {
        "type": "object",
        "properties": {
          "itemId": {
            "type": "number"
          }
        }
      },
      "UpdateItemRequest": {
        "type": "object",
        "properties": {
          "itemId": {
            "type": "number"
          },
          "itemName": {
            "type": "string"
          },
          "itemAmount": {
            "type": "integer"
          },
          "itemPrice": {
            "type": "number"
          }
        }
      },
      "SelectItemListRequest": {
        "type": "object",
        "properties": {}
      },
      "SelectItemByNameRequest": {
        "type": "object",
        "properties": {
          "itemName": {
            "type": "string"
          }
        }
      },
      "SelectItemByIdRequest": {
        "type": "object",
        "properties": {
          "itemId": {
            "type": "number"
          }
        }
      },
      "InsertMemberRequest": {
        "type": "object",
        "properties": {
          "memberType": {
            "type": "integer"
          },
          "memberName": {
            "type": "string"
          },
          "memberPhone": {
            "type": "string"
          }
        }
      },
      "SelectMemberListRequest": {
        "type": "object",
        "properties": {}
      },
      "SelectMemberByIdRequest": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "number"
          }
        }
      },
      "InsertOrderRequest": {
        "type": "object",
        "properties": {
          "orderStatus": {
            "type": "string"
          },
          "memberId": {
            "type": "number"
          },
          "deliveryId": {
            "type": "number"
          }
        }
      },
      "UpdateOrderStatusRequest": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "number"
          }
        }
      },
      "InsertOrderItemRequest": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "number"
          },
          "itemId": {
            "type": "number"
          },
          "itemPrice": {
            "type": "number"
          },
          "itemAmount": {
            "type": "integer"
          },
          "orderId": {
            "type": "number"
          }
        }
      },
      "DeleteOneOrderItemRequest": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "number"
          },
          "itemId": {
            "type": "number"
          }
        }
      },
      "GetOrderitemByMemberIdRequest": {
        "type": "object",
        "properties": {
          "member_id": {
            "type": "number"
          }
        }
      },
      "InsertDeliveryResponse": {
        "type": "object",
        "properties": {
          "deliveryId": {
            "type": "number"
          }
        }
      },
      "UpdateDeliveryResponse": {
        "type": "object",
        "properties": {
          "deliveryId": {
            "type": "number"
          }
        }
      },
      "InsertItemResponse": {
        "type": "object",
        "properties": {
          "itemId": {
            "type": "number"
          }
        }
      },
      "DeleteItemResponse": {
        "type": "object",
        "properties": {
          "itemId": {
            "type": "number"
          }
        }
      },
      "UpdateItemResponse": {
        "type": "object",
        "properties": {
          "itemId": {
            "type": "number"
          }
        }
      },
      "SelectItemListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          }
        }
      },
      "SelectItemByNameResponse": {
        "type": "object",
        "properties": {
          "item": {
            "$ref": "#/components/schemas/Item"
          }
        }
      },
      "SelectItemByIdResponse": {
        "type": "object",
        "properties": {
          "item": {
            "$ref": "#/components/schemas/Item"
          }
        }
      },
      "InsertMemberResponse": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "number"
          }
        }
      },
      "SelectMemberListResponse": {
        "type": "object",
        "properties": {
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Member"
            }
          }
        }
      },
      "SelectMemberByIdResponse": {
        "type": "object",
        "properties": {
          "member": {
            "$ref": "#/components/schemas/Member"
          }
        }
      },
      "InsertOrderResponse": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "number"
          }
        }
      },
      "UpdateOrderStatusResponse": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "number"
          }
        }
      },
      "InsertOrderItemResponse": {
        "type": "object",
        "properties": {
          "orderitemId": {
            "type": "number"
          }
        }
      },
      "DeleteOneOrderItemResponse": {
        "type": "object",
        "properties": {
          "orderitemId": {
            "type": "number"
          }
        }
      },
      "GetOrderitemByMemberIdResponse": {
        "type": "object",
        "properties": {
          "orderitems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Orderitem"
            }
          }
        }
      }
    }
  }
}