# miniminishop/src/main/java/com/example/miniminishop/controller/request/SelectMemberListRequest.java
package com.example.miniminishop.controller.request;

import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiParam;
import lombok.Data;
import org.springframework.web.multipart.MultipartFile;
import java.util.List;

@ApiModel(description = "회원 리스트 가져오기")
@Data
public class SelectMemberListRequest {
}