# miniminishop/src/main/java/com/example/miniminishop/controller/response/InsertItemResponse.java
package com.example.miniminishop.controller.response;

import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiParam;
import lombok.Data;
import lombok.EqualsAndHashCode;
import java.util.HashMap;
import java.util.List;

@EqualsAndHashCode(callSuper = true)
@ApiModel(description = "상품 추가")
@Data
public class InsertItemResponse extends BaseResponse {

  public long itemId = 0;
}