# miniminishop/src/main/java/com/example/miniminishop/controller/response/SelectMemberByIdResponse.java
package com.example.miniminishop.controller.response;

import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiParam;
import lombok.Data;
import lombok.EqualsAndHashCode;
import com.example.miniminishop.controller.vo.Member;
import java.util.HashMap;
import java.util.List;

@EqualsAndHashCode(callSuper = true)
@ApiModel(description = "회원 번호로 회원 찾아오기")
@Data
public class SelectMemberByIdResponse extends BaseResponse {

  public Member member = null;
}