# miniminishop/src/main/java/com/example/miniminishop/controller/response/UpdateOrderStatusResponse.java
package com.example.miniminishop.controller.response;

import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiParam;
import lombok.Data;
import lombok.EqualsAndHashCode;
import java.util.HashMap;
import java.util.List;

@EqualsAndHashCode(callSuper = true)
@ApiModel(description = "주문 상태 변경")
@Data
public class UpdateOrderStatusResponse extends BaseResponse {

  public long orderId = 0;
}